+++ conda/recipes/cusignal/build.sh
@@ -1,4 +1,3 @@
 #!/bin/bash
 
-cd python
-python setup.py install --single-version-externally-managed --record record.txt
+./build.sh cusignal -c