+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - PR #214 - Fix grid-stride loop bug in polyphase channelizer
 - PR #218 - Remove fatbins from source code on GH
 - PR #221 - Synchronization issue with cusignal testing 
+- PR #237 - Update conda build files so fatbins are generated
 
 # cuSignal 0.15.0 (26 Aug 2020)